[#146]chore: useRef()의 props로 주어질 수 없음을 알리는 경고 해결을 위한 input 추가 스크롤 코드 위치 변경
